+++ eidolon/Sources/SwiftUI/Core.swift
@@ -1,5 +1,6 @@
 import UIKit
 @_exported import CoreGraphics
+@_exported import UIKit
 @_exported import Foundation
 #if REV_SYSTEM_COMBINE || REV_STYX
 @_exported import Combine